# docs/api/openapi_v001.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Managed Identity Wallets API",
    "description": "Managed Identity Wallets API",
    "termsOfService": "https://www.eclipse.org/legal/termsofuse.php",
    "contact": {
      "name": "Eclipse Tractus-X",
      "url": "https://projects.eclipse.org/projects/automotive.tractusx",
      "email": "tractusx-dev@eclipse.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://github.com/eclipse-tractusx/managed-identity-wallets/blob/develop/LICENSE"
    },
    "version": "0.0.1"
  },
  "security": [
    {"Authenticate using access_token": []},
    {"sts_token": []}
  ],
  "servers": [],
  "paths": {
    "/api/wallets": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "List of wallets",
        "description": "Permission: **view_wallets** \n\n Retrieve list of registered wallets",
        "operationId": "getWallets",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number, Page number start with zero",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Number of records per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 2147483647
            }
          },
          {
            "name": "sortColumn",
            "in": "query",
            "description": "Sort column name",
            "required": false,
            "schema": {
              "type": "string",
              "default": "createdAt"
            },
            "examples": {
              "Creation date": {
                "description": "Creation date",
                "value": "createdAt"
              },
              "Wallet BPN": {
                "description": "Wallet BPN",
                "value": "bpn"
              },
              "Wallet did": {
                "description": "Wallet did",
                "value": "did"
              },
              "Wallet name": {
                "description": "Wallet name",
                "value": "name"
              }
            }
          },
          {
            "name": "sortTpe",
            "in": "query",
            "description": "Sort order",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc"
            },
            "examples": {
              "Ascending order": {
                "description": "Ascending order",
                "value": "asc"
              },
              "Descending order": {
                "description": "Descending order",
                "value": "desc"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet list",
            "content": {
              "application/json": {
                "examples": {
                  "Wallet list": {
                    "description": "Wallet list",
                    "value": {
                      "content": [
                        {
                          "name": "companyA",
                          "did": "did:web:localhost:BPNL000000000001",
                          "bpn": "BPNL000000000001",
                          "algorithm": "ED25519",
                          "didDocument": {
                            "@context": [
                              "https://www.w3.org/ns/did/v1",
                              "https://w3c.github.io/vc-jws-2020/contexts/v1"
                            ],
                            "id": "did:web:localhost:BPNL000000000001",
                            "verificationMethod": [
                              {
                                "controller": "did:web:localhost:BPNL000000000001",
                                "id": "did:web:localhost:BPNL000000000001#",
                                "publicKeyJwk": {
                                  "crv": "Ed25519",
                                  "kty": "OKP",
                                  "x": "mhph0ZSVk7cDVmazbaaC3jBDpphW4eNygAK9gHPlMow"
                                },
                                "type": "JsonWebKey2020"
                              }
                            ]
                          }
                        }
                      ],
                      "pageable": {
                        "sort": {
                          "empty": false,
                          "sorted": true,
                          "unsorted": false
                        },
                        "offset": 0,
                        "pageNumber": 0,
                        "pageSize": 1,
                        "paged": true,
                        "unpaged": false
                      },
                      "totalElements": 3,
                      "totalPages": 3,
                      "last": false,
                      "size": 1,
                      "number": 0,
                      "sort": {
                        "empty": false,
                        "sorted": true,
                        "unsorted": false
                      },
                      "first": true,
                      "numberOfElements": 1,
                      "empty": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input does not comply to the syntax requirements",
            "content": {
              "application/json": {
                "examples": {
                  "Response in case of invalid data provided": {
                    "description": "Response in case of invalid data provided",
                    "value": {
                      "type": "about:blank",
                      "title": "title",
                      "status": 400,
                      "detail": "details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689760833962,
                        "errors": {}
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed due to a failed authorization."
          },
          "403": {
            "description": "The request could not be completed due to a forbidden access"
          },
          "500": {
            "description": "Any other internal server error",
            "content": {
              "application/json": {
                "examples": {
                  "Internal server error": {
                    "description": "Internal server error",
                    "value": {
                      "type": "about:blank",
                      "title": "Error Title",
                      "status": 500,
                      "detail": "Error Details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689762476720
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authenticate using access_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "Wallets"
        ],
        "summary": "Create Wallet",
        "description": "Permission: **add_wallets** (The BPN of the base wallet must equal BPN of caller)\n\n Create a wallet and store it",
        "operationId": "createWallet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWalletRequest"
              },
              "examples": {
                "Create wallet with BPN": {
                  "description": "Create wallet with BPN",
                  "value": {
                    "businessPartnerNumber": "BPNL000000000001",
                    "companyName": "companyA",
                    "didUrl": "portal.com:BPNL000000000001"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "examples": {
                  "Success response": {
                    "description": "Success response",
                    "value": {
                      "name": "companyA",
                      "did": "did:web:localhost:BPNL000000000001",
                      "bpn": "BPNL000000000501",
                      "algorithm": "ED25519",
                      "didDocument": {
                        "@context": [
                          "https://www.w3.org/ns/did/v1",
                          "https://w3c.github.io/vc-jws-2020/contexts/v1"
                        ],
                        "id": "did:web:localhost:BPNL000000000001",
                        "verificationMethod": [
                          {
                            "controller": "did:web:localhost:BPNL000000000001",
                            "id": "did:web:localhost:BPNL000000000001#key-1",
                            "publicKeyJwk": {
                              "crv": "Ed25519",
                              "kty": "OKP",
                              "x": "0Ap6FsX5UuRBIoOzxWtcFA2ymnqXw0U08Ino_mIuYM4"
                            },
                            "type": "JsonWebKey2020"
                          },
                          {
                            "controller": "did:web:localhost:BPNL000000000001",
                            "id": "did:web:localhost:BPNL000000000001#key-2",
                            "publicKeyJwk": {
                              "crv": "secp256k1",
                              "kty": "EC",
                              "x": "f9PkTOpsbcgKe_-s6bNCve3-aB1VZAFsCub8C5bhDn0",
                              "y": "xH1d7jCFavolGVZtaWcZZGP2nLuEsamDCotD56llxUk"
                            },
                            "type": "JsonWebKey2020"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input does not comply to the syntax requirements",
            "content": {
              "application/json": {
                "examples": {
                  "Response in case of invalid data provided": {
                    "description": "Response in case of invalid data provided",
                    "value": {
                      "type": "about:blank",
                      "title": "Invalid data provided",
                      "status": 400,
                      "detail": "details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689760833962,
                        "errors": {
                          "filed": "filed error message"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed due to a failed authorization."
          },
          "403": {
            "description": "The request could not be completed due to a forbidden access"
          },
          "409": {
            "description": "The request could not be completed due to a conflict.",
            "content": {
              "application/json": {
                "examples": {
                  "Wallet already exist": {
                    "description": "Wallet already exist",
                    "value": {
                      "type": "about:blank",
                      "title": "Wallet is already exists for bpn BPNL000000000001",
                      "status": 409,
                      "detail": "Wallet is already exists for bpn BPNL000000000001",
                      "instance": "/api/wallets",
                      "properties": {
                        "timestamp": 1689762639948
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Any other internal server error",
            "content": {
              "application/json": {
                "examples": {
                  "Internal server error": {
                    "description": "Internal server error",
                    "value": {
                      "type": "about:blank",
                      "title": "Error Title",
                      "status": 500,
                      "detail": "Error Details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689762476720
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authenticate using access_token": []
          }
        ]
      }
    },
    "/api/wallets/{identifier}/credentials": {
      "post": {
        "tags": [
          "Wallets"
        ],
        "summary": "Store Verifiable Credential",
        "description": "Permission: **update_wallets** OR **update_wallet** (The BPN of wallet to extract credentials from must equal BPN of caller) \n\n Store a verifiable credential in the wallet of the given identifier",
        "operationId": "storeCredential",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "description": "Did or BPN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "examples": {
              "bpn": {
                "description": "bpn",
                "value": "BPNL000000000000"
              },
              "did": {
                "description": "did",
                "value": "did:web:localhost:BPNL000000000000"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "object"
                }
              },
              "example": {
                "@context": [
                  "https://www.w3.org/2018/credentials/v1",
                  "https://registry.lab.gaia-x.eu/development/api/trusted-shape-registry/v1/shapes/jsonld/trustframework#"
                ],
                "id": "did:web:localhost.in#123456789",
                "type": [
                  "VerifiableCredential",
                  "LegalParticipant"
                ],
                "issuer": "did:web:localhost.in",
                "issuanceDate": "2023-05-04T07:36:03.633Z",
                "credentialSubject": {
                  "id": "https://localhost/.well-known/participant.json",
                  "type": "gx:LegalParticipant",
                  "gx:legalName": "Sample Company",
                  "gx:legalRegistrationNumber": {
                    "gx:taxID": "113123123"
                  },
                  "gx:headquarterAddress": {
                    "gx:countrySubdivisionCode": "BE-BRU"
                  },
                  "gx:legalAddress": {
                    "gx:countrySubdivisionCode": "BE-BRU"
                  },
                  "gx-terms-and-conditions:gaiaxTermsAndConditions": "70c1d713215f95191a11d38fe2341faed27d19e083917bc8732ca4fea4976700"
                },
                "proof": {
                  "type": "JsonWebSignature2020",
                  "created": "2023-05-04T07:36:04.079Z",
                  "proofPurpose": "assertionMethod",
                  "verificationMethod": "did:web:localhost",
                  "jws": "eyJhbGciOiJQUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..iHki8WC3nPfcSRkC_AV4tXh0ikfT7BLPTGc_0ecI8zontTmJLqwcpPfAt0PFsoo3SkZgc6j636z55jj5tagBc-OKoiDu7diWryNAnL9ASsmWJyrPhOKVARs6x6PxVaTFBuyCfAHZeipxmkcYfNB_jooIXO2HuRcL2odhsQHELkGc5IDD-aBMWyNpfVAaYQ-cCzvDflZQlsowziUKfMkBfwpwgMdXFIgKWYdDIRvzA-U-XiC11-6QV7tPeKsMguEU0F5bh8cCEm2rooqXtENcsM_7cqFdQoOyblJyM-agoz2LUTj9QIdn9_gnNkGN-2U7_qBJWmHkK1Hm_mHqcNeeQw"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "examples": {
                  "Success Response": {
                    "description": "Success Response",
                    "value": {
                      "message": "Credential with id did:web:localhost#123456789 has been successfully stored"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The input does not comply to the syntax requirements",
            "content": {
              "application/json": {
                "examples": {
                  "Response in case of invalid data provided": {
                    "description": "Response in case of invalid data provided",
                    "value": {
                      "type": "about:blank",
                      "title": "title",
                      "status": 400,
                      "detail": "details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689760833962,
                        "errors": {}
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "The request could not be completed due to a failed authorization."
          },
          "403": {
            "description": "The request could not be completed due to a forbidden access"
          },
          "404": {
            "description": "Wallet not found with provided identifier",
            "content": {
              "application/json": {
                "examples": {
                  "Wallet not found with provided identifier": {
                    "description": "Wallet not found with provided identifier",
                    "value": {
                      "type": "about:blank",
                      "title": "Wallet not found for identifier did:web:localhost:BPNL000000044001",
                      "status": 404,
                      "detail": "Wallet not found for identifier did:web:localhost:BPNL000000044001",
                      "instance": "/api/wallets/did%3Aweb%3Alocalhost%3ABPNL0000000/credentials",
                      "properties": {
                        "timestamp": 1689765541959
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Any other internal server error",
            "content": {
              "application/json": {
                "examples": {
                  "Internal server error": {
                    "description": "Internal server error",
                    "value": {
                      "type": "about:blank",
                      "title": "Error Title",
                      "status": 500,
                      "detail": "Error Details",
                      "instance": "API endpoint",
                      "properties": {
                        "timestamp": 1689762476720
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "Authenticate using access_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateWalletRequest": {
        "required": [
          "businessPartnerNumber",
          "companyName",
          "didUrl"
        ],
        "type": "object",
        "properties": {
          "businessPartnerNumber": {
            "pattern": "^(BPN)(L|S|A)[0-9A-Z]{12}",
            "type": "string"
          },
          "companyName": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "didUrl": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Authenticate using access_token": {
        "type": "apiKey",
        "description": "**Bearer (apiKey)**\nJWT Authorization header using the Bearer scheme.\nEnter **Bearer** [space] and then your token in the text input below:\nExample: Bearer 12345abcdef\n",
        "name": "Authorization",
        "in": "header"
      },
      "sts_token": {
        "type": "apiKey",
        "description": "**STS token**\nJWT Authorization header.\nEnter your token in the text input below:\nExample: 12345abcdef\n",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}